letsema-backend/requirements.txt
django==3.2
djongo
pymongo
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
psycopg2-binary==2.9.9
